Removes CoC violation slide
This commit removes a slide with sexual imagery from Thomas Gelf's
Ignite presentation.

This slide was not presented at Ignite, but the linked PDF
came up on a thread on the puppet-users mailing list and
pretty clearly (most amazingly, with a degree of self-awareness)
is a COC violation:

> Remember that sexual language and imagery is not appropriate for any event venue, including talks.

I've communicated separately with the author.

ref: https://groups.google.com/d/msg/puppet-dev/ve0IdegWQck/H4SisfZeBAAJ