Moves the anchor tags around on the schedule page
I know this doesn't seem to make any sense.  Why does
`<a name="chef"></a>` appear in the middle of the CFEngine talks?!

Well, it's to account for the header on the page so when you click the
"chef" link at the top of the page your browser jumps down to the Chef
section of the page, not just past it.

Signed-off-by: Nathen Harvey <nharvey@chef.io>